+++ src/main/resources/regularNode.conf
@@ -5,7 +5,7 @@ akka {
     artery {
       transport = tcp # See Selecting a transport below
       canonical.hostname = "127.0.0.1" #reachable IP of this host
-      canonical.port = 65010 #the port the actor system should listen on, set to 0 to have it chosen automatically
+      canonical.port = 0 #the port the actor system should listen on, set to 0 to have it chosen automatically
     }
   }
 }
